Move Cosmos packet components to packet:: module
