+++ crates/cosmos/cosmos-client-components/src/components/client.rs
@@ -2,9 +2,7 @@ use cgp_core::prelude::delegate_components;
 use hermes_relayer_components::chain::traits::message_builders::ack_packet::AckPacketMessageBuilderComponent;
 use hermes_relayer_components::chain::traits::message_builders::channel_handshake::ChannelHandshakeMessageBuilderComponent;
 use hermes_relayer_components::chain::traits::message_builders::receive_packet::ReceivePacketMessageBuilderComponent;
-use hermes_relayer_components::chain::traits::message_builders::timeout_unordered_packet::{
-    TimeoutUnorderedPacketMessageBuilderComponent, TimeoutUnorderedPacketPayloadBuilderComponent,
-};
+use hermes_relayer_components::chain::traits::message_builders::timeout_unordered_packet::TimeoutUnorderedPacketMessageBuilderComponent;
 use hermes_relayer_components::chain::traits::message_builders::update_client::UpdateClientMessageBuilderComponent;
 use hermes_relayer_components::chain::traits::packet::fields::PacketFieldsReaderComponent;
 use hermes_relayer_components::chain::traits::packet::from_write_ack::PacketFromWriteAckBuilderComponent;
@@ -13,6 +11,7 @@ use hermes_relayer_components::chain::traits::payload_builders::channel_handshak
 use hermes_relayer_components::chain::traits::payload_builders::connection_handshake::ConnectionHandshakePayloadBuilderComponent;
 use hermes_relayer_components::chain::traits::payload_builders::create_client::CreateClientPayloadBuilderComponent;
 use hermes_relayer_components::chain::traits::payload_builders::receive_packet::ReceivePacketPayloadBuilderComponent;
+use hermes_relayer_components::chain::traits::payload_builders::timeout_unordered_packet::TimeoutUnorderedPacketPayloadBuilderComponent;
 use hermes_relayer_components::chain::traits::payload_builders::update_client::UpdateClientPayloadBuilderComponent;
 use hermes_relayer_components::chain::traits::queries::block::BlockQuerierComponent;
 use hermes_relayer_components::chain::traits::queries::chain_status::ChainStatusQuerierComponent;
@@ -53,8 +52,6 @@ use hermes_relayer_components::chain::traits::types::status::ChainStatusTypeComp
 use hermes_relayer_components::chain::traits::types::timestamp::TimestampTypeComponent;
 use hermes_relayer_components::chain::traits::types::update_client::UpdateClientPayloadTypeComponent;
 
-use crate::impls::ack_packet_message::BuildCosmosAckPacketMessage;
-use crate::impls::ack_packet_payload::BuildCosmosAckPacketPayload;
 use crate::impls::channel::channel_handshake_message::BuildCosmosChannelHandshakeMessage;
 use crate::impls::channel::channel_handshake_payload::BuildCosmosChannelHandshakePayload;
 use crate::impls::channel::init_channel_options::ProvideCosmosInitChannelOptionsType;
@@ -63,8 +60,14 @@ use crate::impls::client::update_client_message::BuildCosmosUpdateClientMessage;
 use crate::impls::client::update_client_payload::BuildUpdateClientPayloadWithChainHandle;
 use crate::impls::connection::connection_handshake_payload::BuildCosmosConnectionHandshakePayload;
 use crate::impls::connection::init_connection_options::ProvideCosmosInitConnectionOptionsType;
-use crate::impls::packet_fields::CosmosPacketFieldReader;
-use crate::impls::packet_from_ack::BuildCosmosPacketFromWriteAck;
+use crate::impls::packet::ack_packet_message::BuildCosmosAckPacketMessage;
+use crate::impls::packet::ack_packet_payload::BuildCosmosAckPacketPayload;
+use crate::impls::packet::packet_fields::CosmosPacketFieldReader;
+use crate::impls::packet::packet_from_ack::BuildCosmosPacketFromWriteAck;
+use crate::impls::packet::receive_packet_message::BuildCosmosReceivePacketMessage;
+use crate::impls::packet::receive_packet_payload::BuildCosmosReceivePacketPayload;
+use crate::impls::packet::timeout_packet_message::BuildCosmosTimeoutPacketMessage;
+use crate::impls::packet::timeout_packet_payload::BuildCosmosTimeoutPacketPayload;
 use crate::impls::queries::block::QueryCometBlock;
 use crate::impls::queries::chain_id::QueryChainIdWithChainHandle;
 use crate::impls::queries::chain_status::QueryChainStatusWithChainHandle;
@@ -75,11 +78,7 @@ use crate::impls::queries::send_packet::QueryCosmosSendPacket;
 use crate::impls::queries::send_packets::QuerySendPacketsConcurrently;
 use crate::impls::queries::unreceived_packet::QueryUnreceivedCosmosPacketSequences;
 use crate::impls::queries::write_ack_event::QueryWriteAckEventFromChainHandle;
-use crate::impls::receive_packet_message::BuildCosmosReceivePacketMessage;
-use crate::impls::receive_packet_payload::BuildCosmosReceivePacketPayload;
 use crate::impls::send_messages_as_tx::SendMessagesToTxContext;
-use crate::impls::timeout_packet_message::BuildCosmosTimeoutPacketMessage;
-use crate::impls::timeout_packet_payload::BuildCosmosTimeoutPacketPayload;
 use crate::impls::types::chain::ProvideCosmosChainTypes;
 use crate::impls::types::create_client_options::ProvideCosmosCreateClientSettings;
 use crate::impls::types::payload::ProvideCosmosPayloadTypes;
