+++ crates/cosmos/cosmos-client-components/src/impls/mod.rs
@@ -1,14 +1,7 @@
-pub mod ack_packet_message;
-pub mod ack_packet_payload;
 pub mod channel;
 pub mod client;
 pub mod connection;
-pub mod packet_fields;
-pub mod packet_from_ack;
+pub mod packet;
 pub mod queries;
-pub mod receive_packet_message;
-pub mod receive_packet_payload;
 pub mod send_messages_as_tx;
-pub mod timeout_packet_message;
-pub mod timeout_packet_payload;
 pub mod types;
