+++ crates/relayer/relayer-components/src/chain/traits/payload_builders/mod.rs
@@ -3,4 +3,5 @@ pub mod channel_handshake;
 pub mod connection_handshake;
 pub mod create_client;
 pub mod receive_packet;
+pub mod timeout_unordered_packet;
 pub mod update_client;
