+++ crates/relayer/relayer-components/src/relay/components/packet_relayers/timeout_unordered/timeout_unordered_packet.rs
@@ -1,8 +1,7 @@
 use cgp_core::async_trait;
 
-use crate::chain::traits::message_builders::timeout_unordered_packet::{
-    CanBuildTimeoutUnorderedPacketMessage, CanBuildTimeoutUnorderedPacketPayload,
-};
+use crate::chain::traits::message_builders::timeout_unordered_packet::CanBuildTimeoutUnorderedPacketMessage;
+use crate::chain::traits::payload_builders::timeout_unordered_packet::CanBuildTimeoutUnorderedPacketPayload;
 use crate::chain::traits::queries::client_state::CanQueryClientState;
 use crate::chain::types::aliases::HeightOf;
 use crate::relay::traits::chains::{CanRaiseRelayChainErrors, HasRelayChains};
