+++ crates/solomachine/solomachine-relayer/src/impls/chain/solomachine_components/timeout_packet_payload.rs
@@ -1,6 +1,6 @@
 use cgp_core::prelude::*;
 use hermes_cosmos_client_components::methods::encode::encode_protobuf;
-use hermes_relayer_components::chain::traits::message_builders::timeout_unordered_packet::TimeoutUnorderedPacketPayloadBuilder;
+use hermes_relayer_components::chain::traits::payload_builders::timeout_unordered_packet::TimeoutUnorderedPacketPayloadBuilder;
 use ibc_relayer_types::core::ics04_channel::packet::Packet;
 use ibc_relayer_types::core::ics24_host::path::CommitmentsPath;
 use ibc_relayer_types::Height;
